exercio 41 reescrito codigo - treino
